[TWINFACES-427] renamed entities/option to entities/datalist-option
