+++ src/entities/datalist-option/components/resource-link/index.ts
