+++ src/entities/datalist-option/components/resource-link/resource-link.tsx
