+++ src/entities/datalist-option/libs/helpers.ts
